rename python module to src as it might be confusing
