+++ demo.py
@@ -1,13 +1,12 @@
-import sys
-sys.path.insert(0, 'python')
 import cv2
-import model
-import util
-from hand import Hand
-from body import Body
 import matplotlib.pyplot as plt
 import copy
 import numpy as np
+
+from src import model
+from src import util
+from src.body import Body
+from src.hand import Hand
 
 body_estimation = Body('model/body_pose_model.pth')
 hand_estimation = Hand('model/hand_pose_model.pth')
