+++ src/body.py
@@ -8,8 +8,8 @@ import matplotlib
 import torch
 from torchvision import transforms
 
-from python import util
-from python.model import bodypose_model
+from src import util
+from src.model import bodypose_model
 
 class Body(object):
     def __init__(self, model_path):
