+++ src/hand_model_output_size.json
