+++ src/hand_model_outputsize.py
@@ -1,7 +1,8 @@
-from model import handpose_model
 import torch
 from tqdm import tqdm
 import json
+
+from src.model import handpose_model
 
 model = handpose_model()
